requirements.txt
anyio==3.6.2
bcrypt==4.1.2
cffi==1.16.0
click==8.1.3
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
fastapi==0.89.1
h11==0.14.0
idna==3.4
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.4
PyJWT==2.8.0
PyMySQL==1.1.0
python-jose==3.3.0
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.1
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
